remove redundant test, move to better location
